refactor: move KnotVector and BasisFunction one folder up
