+++ src/pynurbs/responsive/__init__.py
@@ -1,3 +1,1 @@
 from .curves import Curve
-from .basis_functions import BasisFunctions
-from .knotspace import GeneratorKnotVector, KnotVector
