+++ tests/operations/test_advanced.py
@@ -13,8 +13,8 @@ from pynurbs.responsive.curves import Curve
 @pytest.mark.order(42)
 @pytest.mark.dependency(
     depends=[
-        "tests/responsive/test_knotspace.py::test_end",
-        "tests/responsive/test_basis_functions.py::test_end",
+        "tests/test_knotspace.py::test_end",
+        "tests/test_basis_functions.py::test_end",
         "tests/responsive/test_beziercurve.py::test_end",
         "tests/responsive/test_splinecurve.py::test_end",
         "tests/operations/test_calculus.py::test_end",
