+++ tests/responsive/test_rationalcurve.py
@@ -5,15 +5,15 @@ from fractions import Fraction as frac
 import numpy as np
 import pytest
 
+from pynurbs.knotspace import GeneratorKnotVector
 from pynurbs.responsive.curves import Curve
-from pynurbs.responsive.knotspace import GeneratorKnotVector
 
 
 @pytest.mark.order(35)
 @pytest.mark.dependency(
     depends=[
-        "tests/responsive/test_knotspace.py::test_end",
-        "tests/responsive/test_basis_functions.py::test_end",
+        "tests/test_knotspace.py::test_end",
+        "tests/test_basis_functions.py::test_end",
         "tests/responsive/test_beziercurve.py::test_end",
         "tests/responsive/test_splinecurve.py::test_end",
     ],
