+++ tests/responsive/test_splinecurve.py
@@ -3,15 +3,15 @@ from copy import copy
 import numpy as np
 import pytest
 
+from pynurbs.knotspace import GeneratorKnotVector, KnotVector
 from pynurbs.responsive.curves import Curve
-from pynurbs.responsive.knotspace import GeneratorKnotVector, KnotVector
 
 
 @pytest.mark.order(34)
 @pytest.mark.dependency(
     depends=[
-        "tests/responsive/test_knotspace.py::test_end",
-        "tests/responsive/test_basis_functions.py::test_end",
+        "tests/test_knotspace.py::test_end",
+        "tests/test_basis_functions.py::test_end",
         "tests/responsive/test_beziercurve.py::test_end",
     ],
     scope="session",
